test(html-player): cover shared-controls source ownership
